+++ tests/ui/real_time_battery_display_tests.rs
@@ -11,7 +11,7 @@ fn test_real_time_battery_updates() {
     let levels = [100, 95, 90, 85, 80, 75, 70];
     
     for &level in &levels {
-        let _widget = view_circular_battery_widget(level, false);
+        let _widget = view_circular_battery_widget(level as f32, false);
         let _icon = battery_icon_display(Some(level), false, 80.0, 20.0);
     }
 }
@@ -22,12 +22,12 @@ fn test_charging_state_transitions() {
     let level = 50;
     
     // Simulate transition from not charging to charging
-    let _not_charging = view_circular_battery_widget(level, false);
-    let _charging = view_circular_battery_widget(level, true);
+    let _not_charging = view_circular_battery_widget(level as f32, false);
+    let _charging = view_circular_battery_widget(level as f32, true);
     
     // Test with different levels during charging
     for charge_level in 50..=100 {
-        let _charging_widget = view_circular_battery_widget(charge_level, true);
+        let _charging_widget = view_circular_battery_widget(charge_level as f32, true);
     }
 }
 
@@ -38,7 +38,7 @@ fn test_rapid_battery_changes() {
     for i in 0..100 {
         let level = (50 + (i % 50)) as u8;
         let charging = i % 10 < 5; // Alternating charging state
-        let _widget = view_circular_battery_widget(level, charging);
+        let _widget = view_circular_battery_widget(level as f32, charging);
     }
 }
 
@@ -53,7 +53,7 @@ fn test_battery_level_smoothing() {
     for i in 0..=steps {
         let progress = i as f32 / steps as f32;
         let current_level = start_level + ((end_level - start_level) as f32 * progress) as u8;
-        let _widget = view_circular_battery_widget(current_level, false);
+        let _widget = view_circular_battery_widget(current_level as f32, false);
     }
 }
 
@@ -64,8 +64,8 @@ fn test_low_battery_warnings() {
     let low_levels = [15, 10, 5, 3, 1, 0];
     
     for &level in &low_levels {
-        let _widget = view_circular_battery_widget(level, false);
-        let _widget_charging = view_circular_battery_widget(level, true);
+        let _widget = view_circular_battery_widget(level as f32, false);
+        let _widget_charging = view_circular_battery_widget(level as f32, true);
     }
 }
 
@@ -76,8 +76,8 @@ fn test_critical_battery_levels() {
     let critical_levels = [0, 1, 2, 3, 4, 5];
     
     for &level in &critical_levels {
-        let _critical_widget = view_circular_battery_widget(level, false);
-        let _critical_charging = view_circular_battery_widget(level, true);
+        let _critical_widget = view_circular_battery_widget(level as f32, false);
+        let _critical_charging = view_circular_battery_widget(level as f32, true);
         
         // Test corresponding icons
         let _critical_icon = battery_icon_display(Some(level), false, 80.0, 20.0);
@@ -92,8 +92,8 @@ fn test_full_charge_scenarios() {
     let full_levels = [95, 96, 97, 98, 99, 100];
     
     for &level in &full_levels {
-        let _full_widget = view_circular_battery_widget(level, false);
-        let _full_charging = view_circular_battery_widget(level, true);
+        let _full_widget = view_circular_battery_widget(level as f32, false);
+        let _full_charging = view_circular_battery_widget(level as f32, true);
     }
 }
 
@@ -108,9 +108,9 @@ fn test_multiple_device_displays() {
     ];
     
     for &(level, charging) in &devices {
-        let _left_widget = view_circular_battery_widget(level, charging);
-        let _right_widget = view_circular_battery_widget(level + 5, charging);
-        let _case_widget = view_circular_battery_widget(level + 10, charging);
+        let _left_widget = view_circular_battery_widget(level as f32, charging);
+        let _right_widget = view_circular_battery_widget((level + 5) as f32, charging);
+        let _case_widget = view_circular_battery_widget((level + 10) as f32, charging);
     }
 }
 
@@ -121,7 +121,7 @@ fn test_display_performance() {
     for cycle in 0..50 {
         for level in 0..=100 {
             let charging = (cycle + level) % 3 == 0;
-            let _widget = view_circular_battery_widget(level, charging);
+            let _widget = view_circular_battery_widget(level as f32, charging);
         }
     }
 }
@@ -138,12 +138,12 @@ fn test_asymmetric_battery_levels() {
     ];
     
     for &(left, right) in &asymmetric_cases {
-        let _left_widget = view_circular_battery_widget(left, false);
-        let _right_widget = view_circular_battery_widget(right, false);
+        let _left_widget = view_circular_battery_widget(left as f32, false);
+        let _right_widget = view_circular_battery_widget(right as f32, false);
         
         // Test with charging
-        let _left_charging = view_circular_battery_widget(left, true);
-        let _right_charging = view_circular_battery_widget(right, true);
+        let _left_charging = view_circular_battery_widget(left as f32, true);
+        let _right_charging = view_circular_battery_widget(right as f32, true);
     }
 }
 
@@ -155,8 +155,8 @@ fn test_edge_case_values() {
     
     for &level in &edge_cases {
         // Test all combinations
-        let _widget_normal = view_circular_battery_widget(level, false);
-        let _widget_charging = view_circular_battery_widget(level, true);
+        let _widget_normal = view_circular_battery_widget(level as f32, false);
+        let _widget_charging = view_circular_battery_widget(level as f32, true);
         
         // Test with different icon sizes
         let _small_icon = battery_icon_display(Some(level), false, 40.0, 15.0);
